SlideBox: rename files
