+++ packages/core/src/SlideBox/Arrow.styled.tsx
