+++ packages/core/src/SlideBox/SlideBox.styled.ts
